CI: Adjust test timing
