+++ tests/conftest.py
@@ -2,6 +2,40 @@
 # This module is part of LorryStream and is released under the MIT License.
 # See LICENSE file for more information.
 
+import pytest
+import sqlalchemy as sa
+
 from lorrystream.util.common import setup_logging
+from tests.testcontainers.cratedb import CrateDBContainer
+
+
+class CrateDBFixture:
+    def __init__(self):
+        self.cratedb = None
+        self.setup()
+
+    def setup(self):
+        self.cratedb = CrateDBContainer("crate/crate:nightly")
+        self.cratedb.start()
+
+    def finalize(self):
+        self.cratedb.stop()
+
+    def reset(self):
+        database_url = self.cratedb.get_connection_url()
+        sa_engine = sa.create_engine(database_url)
+        with sa_engine.connect() as conn:
+            conn.exec_driver_sql("DROP TABLE IF EXISTS testdrive;")
+
+    def get_connection_url(self, *args, **kwargs):
+        return self.cratedb.get_connection_url(*args, **kwargs)
+
+
+@pytest.fixture(scope="function")
+def cratedb():
+    fixture = CrateDBFixture()
+    yield fixture
+    fixture.finalize()
+
 
 setup_logging()
