+++ tests/test_engine.py
@@ -10,7 +10,6 @@ import pytest
 import sqlalchemy as sa
 
 from lorrystream.core import ChannelFactory, Engine
-from tests.testcontainers.cratedb import CrateDBContainer
 
 capmqtt_decode_utf8 = True
 
@@ -31,39 +30,11 @@ class engine_single_shot:
         return self
 
     async def __aexit__(self, exc_type, exc_value, exc_tb):
-        await asyncio.sleep(0.50)
+        await asyncio.sleep(1.50)
         logger.info("Stopping engine")
         self.engine.stop()
+        await asyncio.sleep(0.25)
         return False
-
-
-class CrateDBFixture:
-    def __init__(self):
-        self.cratedb = None
-        self.setup()
-
-    def setup(self):
-        self.cratedb = CrateDBContainer("crate/crate:nightly")
-        self.cratedb.start()
-
-    def finalize(self):
-        self.cratedb.stop()
-
-    def reset(self):
-        database_url = self.cratedb.get_connection_url()
-        sa_engine = sa.create_engine(database_url)
-        with sa_engine.connect() as conn:
-            conn.exec_driver_sql("DROP TABLE IF EXISTS testdrive;")
-
-    def get_connection_url(self, *args, **kwargs):
-        return self.cratedb.get_connection_url(*args, **kwargs)
-
-
-@pytest.fixture(scope="function")
-def cratedb():
-    fixture = CrateDBFixture()
-    yield fixture
-    fixture.finalize()
 
 
 @pytest.mark.asyncio_cooperative
@@ -79,7 +50,6 @@ async def test_dataframe_to_sql(mosquitto, cratedb, capmqtt):
         reading = {"device": "foo", "temperature": 42.42, "humidity": 84.84}
         capmqtt.publish("testdrive/readings", json.dumps(reading))
         capmqtt.publish("testdrive/readings", json.dumps(reading))
-        await asyncio.sleep(0.50)
 
     # Validate data in storage system.
     sa_engine = sa.create_engine(database_url)
